requirements.txt
requests
pandas
google-cloud-bigquery
pyarrow
python-dotenv